requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
python-dotenv
openai
langdetect
requests